rpm check fix

+++ src/main/flight/pid.h
@@ -214,7 +214,7 @@ typedef struct pidProfile_s {
 
     uint8_t dterm_lpf1_dyn_expo;            // set the curve for dynamic dterm lowpass filter
     uint8_t level_race_mode;                // NFE race mode - when true pitch setpoint calculation is gyro based in level mode
-    uint8_t vbat_sag_compensation;          // Reduce motor output by this percentage of the maximum compensation amount
+    int16_t vbat_sag_compensation;          // Reduce motor output by this percentage of the maximum compensation amount
 
     uint8_t simplified_pids_mode;
     uint8_t simplified_master_multiplier;
